Introduce support for new VM based minicap to support Android 12
This commit introduce a new option to use a new minicap
implementation by specifying --screen-grabber or the env
variable SCREEN_GRABBER. The current native minicap is the default
and is refered as 'minicap-bin'. The new minicap which is packaged
in an apk is usable with 'minicap-apk' and will be used as a
fallback.
At the moment, this apk is expected to be included in
minicap-prebuilt node package at this location
minicap-prebuilt/prebuilt/noarch/minicap.apk.

STFService.apk has also been updated to take into account Android 12

Signed-off-by: Crepieux Pierre <pierre.crepieux@orange.com>